+++ flink-connectors/flink-connector-hive/src/main/java/org/apache/flink/connectors/hive/HiveSource.java
@@ -18,7 +18,6 @@
 
 package org.apache.flink.connectors.hive;
 
-import org.apache.flink.api.common.typeinfo.TypeInformation;
 import org.apache.flink.api.java.typeutils.ResultTypeQueryable;
 import org.apache.flink.connector.file.src.AbstractFileSource;
 import org.apache.flink.connector.file.src.reader.BulkFormat;
@@ -27,28 +26,18 @@ import org.apache.flink.connectors.hive.read.HiveSourceSplit;
 import org.apache.flink.core.io.SimpleVersionedSerializer;
 import org.apache.flink.table.catalog.CatalogTable;
 import org.apache.flink.table.data.RowData;
-import org.apache.flink.table.runtime.types.TypeInfoDataTypeConverter;
-import org.apache.flink.table.types.DataType;
+import org.apache.flink.table.types.logical.RowType;
 import org.apache.flink.util.Preconditions;
 
-import org.apache.hadoop.fs.FileSystem;
-import org.apache.hadoop.fs.Path;
-import org.apache.hadoop.hive.metastore.api.StorageDescriptor;
-import org.apache.hadoop.mapred.FileSplit;
-import org.apache.hadoop.mapred.InputFormat;
 import org.apache.hadoop.mapred.JobConf;
-import org.apache.hadoop.util.ReflectionUtils;
 
 import javax.annotation.Nullable;
 
-import java.io.IOException;
-import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.IntStream;
 
 import static org.apache.flink.connector.file.src.FileSource.DEFAULT_SPLIT_ASSIGNER;
 import static org.apache.flink.util.Preconditions.checkNotNull;
-import static org.apache.hadoop.mapreduce.lib.input.FileInputFormat.INPUT_DIR;
 
 /**
  * A unified data source that reads a hive table.
@@ -57,19 +46,16 @@ public class HiveSource extends AbstractFileSource<RowData, HiveSourceSplit> imp
 
 	private static final long serialVersionUID = 1L;
 
-	// DataType of the records to be returned, with projection applied (if any)
-	private final DataType producedDataType;
-
 	HiveSource(
 			JobConf jobConf,
 			CatalogTable catalogTable,
 			List<HiveTablePartition> partitions,
 			int[] projectedFields,
-			long limit,
+			@Nullable Long limit,
 			String hiveVersion,
 			boolean useMapRedReader,
 			boolean isStreamingSource,
-			DataType producedDataType) {
+			RowType producedDataType) {
 		super(
 				new org.apache.flink.core.fs.Path[1],
 				new HiveSourceFileEnumerator.Provider(partitions, new JobConfWrapper(jobConf)),
@@ -77,7 +63,6 @@ public class HiveSource extends AbstractFileSource<RowData, HiveSourceSplit> imp
 				createBulkFormat(new JobConf(jobConf), catalogTable, projectedFields, hiveVersion, producedDataType, useMapRedReader, limit),
 				null);
 		Preconditions.checkArgument(!isStreamingSource, "HiveSource currently only supports bounded mode");
-		this.producedDataType = producedDataType;
 	}
 
 	@Override
@@ -85,19 +70,14 @@ public class HiveSource extends AbstractFileSource<RowData, HiveSourceSplit> imp
 		return HiveSourceSplitSerializer.INSTANCE;
 	}
 
-	@Override
-	public TypeInformation<RowData> getProducedType() {
-		return (TypeInformation<RowData>) TypeInfoDataTypeConverter.fromDataTypeToTypeInfo(producedDataType);
-	}
-
 	private static BulkFormat<RowData, HiveSourceSplit> createBulkFormat(
 			JobConf jobConf,
 			CatalogTable catalogTable,
 			@Nullable int[] projectedFields,
 			String hiveVersion,
-			DataType producedDataType,
+			RowType producedDataType,
 			boolean useMapRedReader,
-			long limit) {
+			Long limit) {
 		checkNotNull(catalogTable, "catalogTable can not be null.");
 		return new HiveBulkFormatAdapter(
 				new JobConfWrapper(jobConf),
@@ -109,60 +89,5 @@ public class HiveSource extends AbstractFileSource<RowData, HiveSourceSplit> imp
 				producedDataType,
 				useMapRedReader,
 				limit);
-	}
-
-	public static List<HiveSourceSplit> createInputSplits(
-			int minNumSplits,
-			List<HiveTablePartition> partitions,
-			JobConf jobConf) throws IOException {
-		List<HiveSourceSplit> hiveSplits = new ArrayList<>();
-		FileSystem fs = null;
-		for (HiveTablePartition partition : partitions) {
-			StorageDescriptor sd = partition.getStorageDescriptor();
-			Path inputPath = new Path(sd.getLocation());
-			if (fs == null) {
-				fs = inputPath.getFileSystem(jobConf);
-			}
-			// it's possible a partition exists in metastore but the data has been removed
-			if (!fs.exists(inputPath)) {
-				continue;
-			}
-			InputFormat format;
-			try {
-				format = (InputFormat)
-						Class.forName(sd.getInputFormat(), true, Thread.currentThread().getContextClassLoader()).newInstance();
-			} catch (Exception e) {
-				throw new FlinkHiveException("Unable to instantiate the hadoop input format", e);
-			}
-			ReflectionUtils.setConf(format, jobConf);
-			jobConf.set(INPUT_DIR, sd.getLocation());
-			//TODO: we should consider how to calculate the splits according to minNumSplits in the future.
-			org.apache.hadoop.mapred.InputSplit[] splitArray = format.getSplits(jobConf, minNumSplits);
-			for (org.apache.hadoop.mapred.InputSplit inputSplit : splitArray) {
-				Preconditions.checkState(inputSplit instanceof FileSplit,
-						"Unsupported InputSplit type: " + inputSplit.getClass().getName());
-				hiveSplits.add(new HiveSourceSplit((FileSplit) inputSplit, partition, null));
-			}
-		}
-
-		return hiveSplits;
-	}
-
-	public static int getNumFiles(List<HiveTablePartition> partitions, JobConf jobConf) throws IOException {
-		int numFiles = 0;
-		FileSystem fs = null;
-		for (HiveTablePartition partition : partitions) {
-			StorageDescriptor sd = partition.getStorageDescriptor();
-			Path inputPath = new Path(sd.getLocation());
-			if (fs == null) {
-				fs = inputPath.getFileSystem(jobConf);
-			}
-			// it's possible a partition exists in metastore but the data has been removed
-			if (!fs.exists(inputPath)) {
-				continue;
-			}
-			numFiles += fs.listStatus(inputPath).length;
-		}
-		return numFiles;
 	}
 }
