+++ flink-connectors/flink-connector-hive/src/test/java/org/apache/flink/table/module/hive/HiveModuleFactoryTest.java
@@ -28,7 +28,6 @@ import java.util.Map;
 
 import static org.junit.Assert.assertEquals;
 
-
 /**
  * Test for {@link HiveModuleFactory}.
  */
